+++ docker-api/docker-api/get_course_infra.go
@@ -1,0 +1,16 @@
+package main
+
+import (
+	"context"
+
+	pb "github.com/centralesupelec/mydocker/docker-api/protobuf"
+)
+
+func (s *server) GetCourseInfra(ctx context.Context, request *pb.CourseInfraRequest) (*pb.CourseInfraResponse, error) {
+	dockerUtils := NewDockerUtils(s.dockerClient)
+	nodeWithCourseIdLabelExists, err := dockerUtils.nodeWithCourseIdLabelExists(ctx, request.GetCourseID())
+	if err != nil {
+		return nil, err
+	}
+	return &pb.CourseInfraResponse{NodeWithCourseIdLabelExists: nodeWithCourseIdLabelExists}, nil
+}